refactor: tham so api

+++ Bookstore.Share/DTO/ThamSoDTO.cs
@@ -1,5 +1,5 @@
 ﻿public class ThamSoDTO
 {
     public string TenThamSo { get; set; } = string.Empty;
-    public decimal GiaTri { get; set; }
+    public int GiaTri { get; set; }
 }
